requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==2.2.19
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
